correct user delete

+++ app.py
@@ -189,19 +189,23 @@ def update(id):
 
 # Delete user
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
     user_to_delete = Users.query.get_or_404(id)
     name = None;
     form = UserForm()
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully !!")
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template("add_user.html", form = form, name = name, our_users= our_users)
-    except:
-        flash("There was a problem deleting the user.... Try again")
-        return render_template("add_user.html", form = form, name = name, our_users= our_users)
+    if id == current_user.id:
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully !!")
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template("add_user.html", form = form, name = name, our_users= our_users)
+        except:
+            flash("There was a problem deleting the user.... Try again")
+            return render_template("add_user.html", form = form, name = name, our_users= our_users)
+    else:
+        flash("Sorry you are not authorized to delete current user")
         
 
 #Password test page
